Tags tree, tag-counts, groups & exp config etc
The container_tags.html now extends containers.html, inheriting the Groups and
Experimenter configuration panel etc. Removes the duplication of center and right
panels between these two pages.
The query to load_experimenters() specifies whether the 'child_counts' of each
experimenter should be 'projects' or 'tags'.
Orphaned data is filtered by experimenter and group in the tree, but not
in the center panel - see #8300. Also see #8259